requirements.txt
google-genai
langchain
langchain-google-vertexai
langchain-community
playwright
google-cloud-aiplatform
langchain-google-vertexai
nest_asyncio
unstructured
langchain_chroma